Remove @semantic-release/exec, adjust how next version is acquired
