+++ .releaserc.json
@@ -72,12 +72,6 @@
       }
     ],
     [
-      "@semantic-release/exec",
-      {
-        "prepareCmd": "NEXT_VERSION=${nextRelease.version} npm run build"
-      }
-    ],
-    [
       "@semantic-release/changelog",
       {
         "changelogFile": "./CHANGELOG.md"
